+++ addons/config/ReadMe.md
@@ -1,0 +1,13 @@
+# TBMod Config
+## Info
+Alle Vanilla Config Anpassungen die nicht Addonspezifisch sind
+
+## Maintainer
+- [Eron](https://github.com/E-for-Eron)
+
+## Developer:
+- [shukari](https://github.com/shukari)
+- [Eric Ruhland](https://github.com/Er1807)
+
+## Contributors:
+[Contributors](https://github.com/TacticalBaconDevs/TBMod/graphs/contributors)
