+++ addons/config/configs/CfgCommands.hpp
@@ -1,0 +1,11 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgCommands
+{
+    allowedHTMLLoadURIs[] += {
+        "http://tacticalbacon.de*",
+        "https://tacticalbacon.de*"
+    };
+};
