+++ addons/config/configs/CfgEditorCategories.hpp
@@ -1,0 +1,11 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgEditorCategories
+{
+    class EdCat_TB_MainCat
+    {
+        displayName = "#TB - MainStuff";
+    };
+};
