+++ addons/config/configs/CfgEditorSubcategories.hpp
@@ -1,15 +1,19 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
 class CfgEditorSubcategories
 {
     class EdSubcat_TB_Spezial
     {
         displayName = "Spezial";
     };
-    
+
     class EdSubcat_TB_FOBs
     {
         displayName = "FOBs";
     };
-    
+
     class EdSubcat_TB_BriefingBereich
     {
         displayName = "BriefingBereich";
