+++ addons/config/configs/CfgMagazineWells.hpp
@@ -1,0 +1,11 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgMagazineWells
+{
+    class UGL_40x36
+    {
+        ADDON[] = {"3Rnd_Smoke_Grenade_shell_precise"};
+    };
+};
