+++ addons/config/configs/cfgScriptPaths.hpp
@@ -1,0 +1,9 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class cfgScriptPaths
+{
+
+};
+
