+++ addons/config/script_component.hpp
@@ -1,0 +1,12 @@
+/*
+    Part of the TBMod ( https://github.com/TacticalBaconDevs/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+#define COMPONENT config
+#include "\x\TBMod\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#include "\x\TBMod\addons\main\script_macros.hpp"
