--- addons/main/configs/CfgEditorCategories.hpp
@@ -1,7 +1,0 @@
-class CfgEditorCategories
-{
-    class EdCat_TB_MainCat
-    {
-        displayName = "#TB - MainStuff";
-    };
-};
