--- addons/main/configs/CfgFactionClasses.hpp
@@ -1,8 +1,0 @@
-class CfgFactionClasses
-{
-    class NO_CATEGORY;
-    class TB_categorie: NO_CATEGORY
-    {
-        displayName = "#TacticalBacon";
-    };
-};
