+++ addons/main/configs/CfgVehicles.hpp
@@ -282,455 +282,4 @@ class CfgVehicles
     ADD_SIGN(OneTeam);
     ADD_SIGN(Rufnamen);
     ADD_SIGN(Trupp);
-
-
-    // ###################### ACE Defuse ######################
-    // class Items_base_F;
-    // class ACE_DefuseObject: Items_base_F
-    // {
-        // class EventHandlers
-        // {
-            // class CBA_Extended_EventHandlers: CBA_Extended_EventHandlers {};
-        // };
-
-        // author = "ACE-Team";
-        // _generalMacro = "ACE_DefuseObject";
-        // displayName = "ACE Defuse Helper";
-        // model = "\A3\Weapons_f\dummyweapon.p3d";
-        // scope = 1;
-
-        // class ACE_Actions
-        // {
-            // class ACE_Defuse
-            // {
-                // distance = 2;
-                // displayName = "Entschärfen";
-                // condition = "";
-                // statement = ";";
-                // exceptions[] = {"isNotSwimming"};
-                // icon = "";
-            // };
-        // };
-    // };
-    // class ACE_DefuseObject_Large: ACE_DefuseObject
-    // {
-        // class ACE_Actions: ACE_Actions
-        // {
-            // class ACE_Defuse: ACE_Defuse
-            // {
-                // distance = 3;
-            // };
-        // };
-    // };
-
-
-    // ###################### TFAR RADIOS leichter ######################
-    #define ADD_OVERRIDE(NAME,PARENT,LOAD,MASSE) class NAME : PARENT \
-        { \
-            maximumLoad = LOAD; \
-            mass = MASSE; \
-        }
-    class TFAR_Bag_Base;
-    ADD_OVERRIDE(TFAR_rt1523g,TFAR_Bag_Base,129,25); // 50, 80 - 5,16 - Russen kleines LR
-    ADD_OVERRIDE(tf_rt1523g,TFAR_rt1523g,129,25);
-    ADD_OVERRIDE(OPXT_aor1_117,tf_rt1523g,129,25);
-
-    ADD_OVERRIDE(TFAR_rt1523g_big,TFAR_rt1523g,160,30); // 160, 160 - 5,3 - USA großes LR
-    ADD_OVERRIDE(tf_rt1523g_big,TFAR_rt1523g_big,160,30);
-    ADD_OVERRIDE(OPXT_aor1_1523,tf_rt1523g_big,160,30);
-
-    ADD_OVERRIDE(TFAR_rt1523g_sage,TFAR_rt1523g,160,30); // 100, 120 - 5,3
-    ADD_OVERRIDE(tf_rt1523g_sage,TFAR_rt1523g_sage,160,30);
-
-    // OPXT_aor1_210: tf_anarc210 - maximumLoad = 280; mass = 50;
-    // OPXT_rf7800m: tf_anarc210 - maximumLoad = 300; mass = 60;
-
-
-    // ###################### ACE-Minedetector ######################
-    // TODO: CUP Abhängigkeit
-    /*#define SET_DETECTABLE(PARENT,CLASSES) class CLASSES : PARENT {ace_minedetector_detectable = 1;}
-
-    class Small_items_NoInteractive;
-    class Items_base_F;
-    class House_Small_F;
-    class House;
-    class Strategic;
-
-    //SET_DETECTABLE(Small_items_NoInteractive, Explosive);
-    SET_DETECTABLE(Items_base_F, Land_MetalBarrel_F);
-    SET_DETECTABLE(House_Small_F, Land_MetalBarrel_empty_F);
-    SET_DETECTABLE(House, Land_popelnice);
-    SET_DETECTABLE(House, Land_Kontejner);
-    SET_DETECTABLE(Strategic, Garbage_can);
-    SET_DETECTABLE(Strategic, Garbage_container);*/
-
-
-    // ###################### Rucksäcke ######################
-    class Bag_Base;
-    class B_AssaultPack_Base : Bag_Base
-    {
-        maximumLoad = 106;
-        mass = 13;
-    };
-
-
-    // ###################### MISC ######################
-    // UGV Stomper
-    class Car_F;
-    class UGV_01_base_F: Car_F
-    {
-        class Turrets;
-    };
-
-    class UGV_01_rcws_base_F: UGV_01_base_F
-    {
-        class Turrets: Turrets
-        {
-            class MainTurret;
-        };
-    };
-
-    class B_UGV_01_rcws_F: UGV_01_rcws_base_F {};
-
-    class TB_UGV_01_rcws_F: B_UGV_01_rcws_F
-    {
-        displayName = "TB UGV Stomper RCWS (Rauch)"; // UGV Stomper RCWS
-        editorCategory = "EdCat_TB_MainCat";
-        editorSubcategory = "EdSubcat_TB_Spezial";
-
-        class Turrets: Turrets
-        {
-            class MainTurret: MainTurret
-            {
-                magazines[] = {"200Rnd_127x99_mag_Tracer_Red","TB_Smoke_Grenades","TB_Smoke_Grenades"};
-            };
-        };
-    };
-
-
-    // ###################### No Uniform ######################
-    class Civilian;
-    class CAManBase;
-    class B_Soldier_diver_base_F;
-    class C_man_1;
-    class I_G_Soldier_F;
-    class I_G_Soldier_lite_F;
-    class I_G_Soldier_SL_F;
-    class I_G_Soldier_TL_F;
-    class I_G_Soldier_AR_F;
-    class I_G_medic_F;
-    class I_G_engineer_F;
-    class I_G_Soldier_exp_F;
-    class I_G_Soldier_GL_F;
-    class I_G_Soldier_M_F;
-    class I_G_Soldier_LAT_F;
-    class I_G_Soldier_A_F;
-    class I_G_officer_F;
-    class I_Soldier_diver_base_F;
-    class O_Soldier_diver_base_F;
-    class I_G_Sharpshooter_F;
-    class B_Soldier_F;
-    class I_C_Soldier_Para_2_F;
-    class I_G_Soldier_LAT2_F;
-    class I_crew_F;
-    class O_R_Gorka_F;
-    class I_L_Soldier_Base_F;
-
-    class Civilian_F: Civilian {
-        modelSides[] = {6};
-    };
-    class SoldierWB: CAManBase {
-        modelSides[] = {6};
-    };
-    class SoldierEB: CAManBase {
-        modelSides[] = {6};
-    };
-    class SoldierGB: CAManBase {
-        modelSides[] = {6};
-    };
-    class B_Soldier_base_F: SoldierWB {
-        modelSides[] = {6};
-    };
-    class B_Soldier_02_f: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class B_Soldier_03_f: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class B_Soldier_04_f: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class B_Soldier_05_f: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class B_RangeMaster_F: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class B_Helipilot_F: B_Soldier_04_f {
-        modelSides[] = {6};
-    };
-    class B_Pilot_F: B_Soldier_05_f {
-        modelSides[] = {6};
-    };
-    class B_helicrew_F: B_Helipilot_F {
-        modelSides[] = {6};
-    };
-    class B_diver_F: B_Soldier_diver_base_F {
-        modelSides[] = {6};
-    };
-    class B_Story_SF_Captain_F: B_Soldier_02_f {
-        modelSides[] = {6};
-    };
-    class B_Story_Protagonist_F: B_Soldier_02_f {
-        modelSides[] = {6};
-    };
-    class b_soldier_survival_F: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class C_man_1_1_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_man_p_fugitive_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_man_p_beggar_F: C_man_p_fugitive_F {
-        modelSides[] = {6};
-    };
-    class C_man_w_worker_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_scientist_F: C_man_w_worker_F {
-        modelSides[] = {6};
-    };
-    class C_man_hunter_1_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_man_pilot_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_journalist_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class I_G_Soldier_base_F: SoldierGB {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_F: I_G_Soldier_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_F: I_G_Soldier_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_lite_F: I_G_Soldier_lite_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_lite_F: I_G_Soldier_lite_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_SL_F: I_G_Soldier_SL_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_SL_F: I_G_Soldier_SL_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_TL_F: I_G_Soldier_TL_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_TL_F: I_G_Soldier_TL_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_AR_F: I_G_Soldier_AR_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_AR_F: I_G_Soldier_AR_F {
-        modelSides[] = {6};
-    };
-    class B_G_medic_F: I_G_medic_F {
-        modelSides[] = {6};
-    };
-    class O_G_medic_F: I_G_medic_F {
-        modelSides[] = {6};
-    };
-    class B_G_engineer_F: I_G_engineer_F {
-        modelSides[] = {6};
-    };
-    class O_G_engineer_F: I_G_engineer_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_exp_F: I_G_Soldier_exp_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_exp_F: I_G_Soldier_exp_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_GL_F: I_G_Soldier_GL_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_GL_F: I_G_Soldier_GL_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_M_F: I_G_Soldier_M_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_M_F: I_G_Soldier_M_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_LAT_F: I_G_Soldier_LAT_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_LAT_F: I_G_Soldier_LAT_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_A_F: I_G_Soldier_A_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_A_F: I_G_Soldier_A_F {
-        modelSides[] = {6};
-    };
-    class B_G_officer_F: I_G_officer_F {
-        modelSides[] = {6};
-    };
-    class O_G_officer_F: I_G_officer_F {
-        modelSides[] = {6};
-    };
-    class I_G_Story_SF_Captain_F: B_G_Soldier_F {
-        modelSides[] = {6};
-    };
-    class I_Soldier_base_F: SoldierGB {
-        modelSides[] = {6};
-    };
-    class I_Soldier_03_F: I_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class I_Soldier_04_F: I_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class I_officer_F: I_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class I_diver_F: I_Soldier_diver_base_F {
-        modelSides[] = {6};
-    };
-    class O_Soldier_base_F: SoldierEB {
-        modelSides[] = {6};
-    };
-    class O_Soldier_02_F: O_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class O_officer_F: O_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class O_diver_F: O_Soldier_diver_base_F {
-        modelSides[] = {6};
-    };
-    class C_Driver_1_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_Marshal_F: B_RangeMaster_F {
-        modelSides[] = {6};
-    };
-    class B_Soldier_VR_F: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class O_Soldier_VR_F: O_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class I_Soldier_VR_F: I_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class C_Soldier_VR_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class B_Protagonist_VR_F: B_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class O_Protagonist_VR_F: O_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class I_Protagonist_VR_F: I_Soldier_base_F {
-        modelSides[] = {6};
-    };
-    class C_Protagonist_VR_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class B_G_Sharpshooter_F: I_G_Sharpshooter_F {
-        modelSides[] = {6};
-    };
-    class O_G_Sharpshooter_F: I_G_Sharpshooter_F {
-        modelSides[] = {6};
-    };
-    class B_Captain_Pettka_F: B_Soldier_02_f {
-        modelSides[] = {6};
-    };
-    class B_Captain_Jay_F: B_Soldier_02_f {
-        modelSides[] = {6};
-    };
-    class I_Captain_Hladas_F: I_officer_F {
-        modelSides[] = {6};
-    };
-    class Underwear_F: B_Soldier_F {
-        modelSides[] = {6};
-    };
-    class B_G_Captain_Ivan_F: I_C_Soldier_Para_2_F {
-        modelSides[] = {6};
-    };
-    class C_man_sport_1_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_man_sport_2_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_man_sport_3_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_Man_casual_1_F: C_man_1 {
-        modelSides[] = {6};
-    };
-    class C_IDAP_Man_Base_F: Civilian_F {
-        modelSides[] = {6};
-    };
-    class C_Story_Mechanic_01_F: Civilian_F {
-        modelSides[] = {6};
-    };
-    class O_A_soldier_base_F: O_officer_F {
-        modelSides[] = {6};
-    };
-    class B_G_Soldier_LAT2_F: I_G_Soldier_LAT2_F {
-        modelSides[] = {6};
-    };
-    class O_G_Soldier_LAT2_F: I_G_Soldier_LAT2_F {
-        modelSides[] = {6};
-    };
-    class I_Story_Crew_F: I_crew_F {
-        modelSides[] = {6};
-    };
-    class C_E_Man_Base_F: Civilian_F {
-        modelSides[] = {6};
-    };
-    class C_E_LooterJacket_01_F: C_E_Man_Base_F {
-        modelSides[] = {6};
-    };
-    class O_R_Gorka_black_F: O_R_Gorka_F {
-        modelSides[] = {6};
-    };
-    class B_CBRN_Man_Base_F: B_Soldier_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_camo_F: I_G_Soldier_LAT_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_deserter_F: I_G_Soldier_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_tshirt_skull_F: I_L_Soldier_Base_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_tshirt_black_F: I_L_Uniform_01_tshirt_skull_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_tshirt_sport_F: I_L_Uniform_01_tshirt_skull_F {
-        modelSides[] = {6};
-    };
-    class I_L_Uniform_01_tshirt_olive_F: I_L_Uniform_01_tshirt_skull_F {
-        modelSides[] = {6};
-    };
 };
